+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 - Added [Fireworks AI](https://inspect.aisi.org.uk/providers.html#fireworks-ai) model provider.
-- OpenAI: Add `http_client` custom model argument.
+- OpenAI: Add `user` and `http_client` custom model arguments.
 - Model API: `--max-connections`, `--max-retries`, and `--timeout` now provide defaults for all models rather than only the main model being evaluated.
 - Datasets: Support for directories in sample `files` field.
 - Added sample, message, and event linking to `log_viewer()` data preparation function.